+++ crates/air_r_formatter/src/r/auxiliary/binary_expression.rs
@@ -6,14 +6,26 @@ use crate::r::auxiliary::call_arguments::FormatRCallArgumentsOptions;
 use air_r_syntax::AnyRExpression;
 use air_r_syntax::RBinaryExpression;
 use air_r_syntax::RBinaryExpressionFields;
+use air_r_syntax::RElseClause;
+use air_r_syntax::RExpressionList;
+use air_r_syntax::RForStatement;
+use air_r_syntax::RFunctionDefinition;
+use air_r_syntax::RIfStatement;
 use air_r_syntax::RLanguage;
+use air_r_syntax::RParenthesizedExpression;
+use air_r_syntax::RRepeatStatement;
 use air_r_syntax::RSyntaxKind;
+use air_r_syntax::RUnaryExpression;
+use air_r_syntax::RWhileStatement;
 use biome_formatter::FormatRuleWithOptions;
 use biome_formatter::format_args;
+use biome_formatter::prelude::dynamic_text;
+use biome_formatter::trivia::format_replaced;
 use biome_formatter::write;
 use biome_rowan::AstNode;
 use biome_rowan::SyntaxResult;
 use biome_rowan::SyntaxToken;
+use settings::AssignmentStyle;
 
 #[derive(Default, Debug, Clone, Copy)]
 pub(crate) enum ChainAlignment {
@@ -213,15 +225,219 @@ fn fmt_binary_assignment(
         }
     });
 
+    let operator_format = format_with(|f: &mut RFormatter| match operator.kind() {
+        RSyntaxKind::ASSIGN => {
+            fmt_left_assignment_operator(node, &operator, LeftAssignmentKind::Arrow, f)
+        }
+        RSyntaxKind::EQUAL => {
+            fmt_left_assignment_operator(node, &operator, LeftAssignmentKind::Equal, f)
+        }
+        _ => operator.format().fmt(f),
+    });
+
     write!(
         f,
         [group(&format_args![
             left.format(),
             space(),
-            operator.format(),
+            operator_format,
             right_format
         ])]
     )
+}
+
+enum LeftAssignmentKind {
+    Arrow,
+    Equal,
+}
+
+fn fmt_left_assignment_operator(
+    node: &RBinaryExpression,
+    operator: &SyntaxToken<RLanguage>,
+    left_assignment_kind: LeftAssignmentKind,
+    f: &mut RFormatter,
+) -> FormatResult<()> {
+    match f.options().assignment_style() {
+        AssignmentStyle::Arrow => match left_assignment_kind {
+            LeftAssignmentKind::Equal => format_replaced(
+                operator,
+                &dynamic_text("<-", operator.text_trimmed_range().start()),
+            )
+            .fmt(f),
+            LeftAssignmentKind::Arrow => {
+                // Already matches style
+                operator.format().fmt(f)
+            }
+        },
+        AssignmentStyle::Equal => match left_assignment_kind {
+            LeftAssignmentKind::Equal => {
+                // Already matches style
+                operator.format().fmt(f)
+            }
+            LeftAssignmentKind::Arrow => {
+                if can_normalize_to_equal(node) {
+                    format_replaced(
+                        operator,
+                        &dynamic_text("=", operator.text_trimmed_range().start()),
+                    )
+                    .fmt(f)
+                } else {
+                    operator.format().fmt(f)
+                }
+            }
+        },
+        AssignmentStyle::Preserve => {
+            // Preserve as is
+            operator.format().fmt(f)
+        }
+    }
+}
+
+/// Can we safely normalize `<-` to `=`?
+///
+/// In R, it is always safe to normalize `=` to `<-`, but the reverse is not true.
+///
+/// For example, it may change the semantic meaning:
+///
+/// ```r
+/// # Expression `x <- 1` as unnamed argument to function `f`
+/// f(x <- 1)
+///
+/// # Named argument `x` with value `1` to function `f`
+/// f(x = 1)
+/// ```
+///
+/// Or it may become a syntax error:
+///
+/// ```r
+/// # Expression `x <- 1`'s result is used as the `condition`
+/// if (x <- 1) this
+///
+/// # Syntax error
+/// if (x = 1) this
+/// ```
+///
+/// To handle this precisely, we looked for all usage of `expr_or_assign_or_help` in the
+/// R `gram.y` grammar. This is the only place that `EQ_ASSIGN` is used in an assignment
+/// context. Each usage of `expr_or_assign_or_help` in `gram.y` is replicated below,
+/// ensuring that we capture all possible places that `=` is allowed.
+/// https://github.com/wch/r-source/blob/b7e27523048d135e3a02560e51cb266702bd49c1/src/main/gram.y#L453-L455
+///
+/// Note how `LEFT_ASSIGN` is instead part of `expr`, making usage of it more permissive
+/// than `EQ_ASSIGN` (a superset, really). This is why we can always replace `=` with
+/// `<-`.
+/// https://github.com/wch/r-source/blob/b7e27523048d135e3a02560e51cb266702bd49c1/src/main/gram.y#L497
+///
+/// `EQ_ASSIGN` is also used in `sub:` and `formlist:`, but these correspond to named
+/// arguments (i.e. `fn(x = 1)`) and parameters with defaults (i.e. `function(x = 1) {}`)
+/// respectively, so are not relevant usages for us (and in fact those are locations where
+/// normalizing to `=` is not allowed).
+/// https://github.com/wch/r-source/blob/b7e27523048d135e3a02560e51cb266702bd49c1/src/main/gram.y#L549-L564
+fn can_normalize_to_equal(node: &RBinaryExpression) -> bool {
+    let node = node.syntax();
+
+    let Some(parent) = node.parent() else {
+        // Should not happen, should always be at least contained in an RRoot's
+        // RExpressionList. We return the conservative `false` if we somehow get here.
+        return false;
+    };
+    let parent_kind = parent.kind();
+
+    // i.e. top level `x <- 1` to `x = 1`
+    // i.e. `{ x <- 1 }` to `{ x = 1 }`
+    if RExpressionList::can_cast(parent_kind) {
+        return true;
+    }
+
+    // i.e. `(x <- 1)` to `(x = 1)`
+    if RParenthesizedExpression::can_cast(parent_kind) {
+        return true;
+    }
+
+    // i.e. `? x <- 1` to `? x = 1`
+    if let Some(parent) = RUnaryExpression::cast_ref(&parent)
+        && let Ok(operator) = parent.operator()
+        && operator.kind() == RSyntaxKind::WAT
+    {
+        return true;
+    }
+
+    // i.e. `x <- y <- 1` to `x = y = 1`
+    // i.e. `x = y <- 1` to `x = y = 1`, but notably `x <- y = 1` is a parse error.
+    // i.e. `x <- 1 ? y` to `x = 1 ? y`
+    // i.e. `y ? x <- 1` to `y ? x = 1`
+    //
+    // We recurse through `can_normalize_to_equal()` in these cases because the parent
+    // binary expression must also be in a position where `=` would be valid. These would
+    // all result in parse errors or change semantic meaning if we didn't check this:
+    //
+    // i.e. `f(x <- y <- 1)` to `f(x <- y = 1)` gives parse error
+    // i.e. `if (x <- y <- 1) z` to `if (x <- y = 1) z` gives parse error
+    //
+    // i.e. `f(x ? y <- 1)` to `f(x ? y = 1)` gives parse error
+    // i.e. `if(x ? y <- 1) z` to `if(x ? y = 1) z` gives parse error
+    //
+    // i.e. `f(x <- 1 ? y)` to `f(x = 1 ? y)` changes semantic meaning
+    // i.e. `if(x <- 1 ? y) z` to `if(x = 1 ? y) z` gives parse error
+    if let Some(parent) = RBinaryExpression::cast_ref(&parent)
+        && let Ok(operator) = parent.operator()
+        && matches!(
+            operator.kind(),
+            RSyntaxKind::ASSIGN | RSyntaxKind::EQUAL | RSyntaxKind::WAT
+        )
+    {
+        return can_normalize_to_equal(&parent);
+    }
+
+    // i.e. `function(x) x <- 1` to `function(x) x = 1`
+    if let Some(parent) = RFunctionDefinition::cast_ref(&parent)
+        && let Ok(body) = parent.body()
+        && body.syntax() == node
+    {
+        return true;
+    }
+
+    // i.e. `if (cond) x <- 1` to `if (cond) x = 1`
+    if let Some(parent) = RIfStatement::cast_ref(&parent)
+        && let Ok(consequence) = parent.consequence()
+        && consequence.syntax() == node
+    {
+        return true;
+    }
+
+    // i.e. `if (cond) x else y <- 1` to `if (cond) x else y = 1`
+    if let Some(parent) = RElseClause::cast_ref(&parent)
+        && let Ok(alternative) = parent.alternative()
+        && alternative.syntax() == node
+    {
+        return true;
+    }
+
+    // i.e. `for(i in 1:5) x <- 1` to `for(i in 1:5) x = 1`
+    if let Some(parent) = RForStatement::cast_ref(&parent)
+        && let Ok(body) = parent.body()
+        && body.syntax() == node
+    {
+        return true;
+    }
+
+    // i.e. `while(cond) x <- 1` to `while(cond) x = 1`
+    if let Some(parent) = RWhileStatement::cast_ref(&parent)
+        && let Ok(body) = parent.body()
+        && body.syntax() == node
+    {
+        return true;
+    }
+
+    // i.e. `repeat x <- 1` to `repeat x = 1`
+    if let Some(parent) = RRepeatStatement::cast_ref(&parent)
+        && let Ok(body) = parent.body()
+        && body.syntax() == node
+    {
+        return true;
+    }
+
+    false
 }
 
 fn format_assignment_rhs(right: &AnyRExpression, table: bool) -> impl Format<RFormatContext> {